Rename test group.
