+++ packages/ckeditor5-paste-from-office/tests/_data/other/index.js
